Crate libra_types

Re-exports

pub use account_address::AccountAddress as PeerId;

Modules

access_path
Suppose we have the following data structure in a smart contract:

account_address

account_config

account_state_blob

block_info

block_metadata

byte_array

contract_event

crypto_proxies
This module defines a data structure made to contain a cryptographic signature, in the sense of an implementation of libra_crypto::traits::Signature. The container is an opaque NewType that intentionally does not allow access to the inner impl.

event

get_with_proof

identifier
An identifier is the name of an entity (module, resource, function, etc) in Move.

language_storage

ledger_info

proof

proptest_types

proto

test_helpers

transaction

validator_change

validator_public_keys

validator_set

validator_signer

validator_verifier

vm_error

write_set
For each transaction the VM executes, the VM will output a WriteSet that contains each access path it updates. For each access path, the VM can either give its new value or delete it.