Module libra_types::account_config

Structs

ACCOUNT_RECEIVED_EVENT_PATH
Returns the path to the received event counter for an Account resource. It can be used to query the event DB for the given event.

ACCOUNT_SENT_EVENT_PATH
The path to the sent event counter for an Account resource. It can be used to query the event DB for the given event.

AccountEvent
Generic struct that represents an Account event. Both SentPaymentEvent and ReceivedPaymentEvent are representable with this struct. They have an AccountAddress for the sender or receiver and the amount transferred.

AccountResource
A Rust representation of an Account resource. This is not how the Account is represented in the VM but it's a convenient representation.

Functions

account_module_name

account_resource_path
Return the path to the Account resource. It can be used to create an AccessPath for an Account resource.

account_struct_name

account_struct_tag

association_address

coin_module_name

coin_struct_name

core_code_address

get_account_resource_or_default

transaction_fee_address

validator_set_address