Struct libra_types::account_config::resources::account::AccountResource

pub struct AccountResource { /* fields omitted */ }

A Rust representation of an Account resource. This is not how the Account is represented in the VM but it's a convenient representation.

Methods

impl AccountResource

pub fn new(
sequence_number: u64,
authentication_key: Vec<u8>,
delegated_key_rotation_capability: bool,
delegated_withdrawal_capability: bool,
sent_events: EventHandle,
received_events: EventHandle,
is_frozen: bool,
balance_currency_code: Identifier
) -> Self

Constructs an Account resource.

pub fn sequence_number(&self) -> u64

Return the sequence_number field for the given AccountResource

pub fn authentication_key(&self) -> &[u8]

Return the authentication_key field for the given AccountResource

pub fn sent_events(&self) -> &EventHandle

Return the sent_events handle for the given AccountResource

pub fn received_events(&self) -> &EventHandle

Return the received_events handle for the given AccountResource

pub fn delegated_key_rotation_capability(&self) -> bool

Return the delegated_key_rotation_capability field for the given AccountResource

pub fn delegated_withdrawal_capability(&self) -> bool

Return the delegated_withdrawal_capability field for the given AccountResource

pub fn is_frozen(&self) -> bool

Return the the is_frozen flag for the given AccountResource

pub fn balance_currency_code(&self) -> &IdentStr

Return the currency code for the currency held by this account