Trait libra_types::on_chain_config::OnChainConfig

pub trait OnChainConfig: Send + Sync + DeserializeOwned {
    const ADDRESS: &'static str;
    const IDENTIFIER: &'static str;
    const CONFIG_ID: ConfigID;
    fn deserialize_default_impl(bytes: &[u8]) -> Result<Self> { ... }

    fn deserialize_into_config(bytes: &[u8]) -> Result<Self> { ... }

    fn fetch_config<T>(storage: T) -> Option<Self>
    where
        T: ConfigStorage,
    { ... }
}

Trait to be implemented by a Rust struct representation of an on-chain config that is stored in storage as a serialized byte array

Associated Constants

const ADDRESS: &'static str

const IDENTIFIER: &'static str

const CONFIG_ID: ConfigID

Loading content...
Provided methods

fn deserialize_default_impl(bytes: &[u8]) -> Result<Self>

fn deserialize_into_config(bytes: &[u8]) -> Result<Self>

fn fetch_config<T>(storage: T) -> Option<Self> where
T: ConfigStorage,

Loading content...