Module libra_types::proof

Re-exports

pub use self::definition::AccountStateProof;

pub use self::definition::AccumulatorConsistencyProof;

pub use self::definition::AccumulatorExtensionProof;

pub use self::definition::AccumulatorProof;

pub use self::definition::AccumulatorRangeProof;

pub use self::definition::EventAccumulatorProof;

pub use self::definition::EventProof;

pub use self::definition::SparseMerkleProof;

pub use self::definition::SparseMerkleRangeProof;

pub use self::definition::TransactionAccumulatorProof;

pub use self::definition::TransactionAccumulatorRangeProof;

pub use self::definition::TransactionInfoWithProof;

pub use self::definition::TransactionListProof;

pub use self::definition::TestAccumulatorProof;

pub use self::definition::TestAccumulatorRangeProof;

Modules

accumulator
This module implements an in-memory Merkle Accumulator that is similar to what we use in storage. This accumulator will only store a small portion of the tree -- for any subtree that is full, we store only the root. Also we only store the frozen nodes, therefore this structure will always store up to Log(n) number of nodes, where n is the total number of leaves in the tree.

definition
This module has definition of various proofs.

position
This module provides an abstraction for positioning a node in a binary tree, A Position uniquely identifies the location of a node

proptest_proof
All proofs generated in this module are not valid proofs. They are only for the purpose of testing conversion between Rust and Protobuf.

Structs

MerkleTreeInternalNode

SparseMerkleLeafNode

SparseMerkleLeafNodeHasher

Type Definitions

EventAccumulatorInternalNode

SparseMerkleInternalNode

TestAccumulatorInternalNode

TransactionAccumulatorInternalNode