Module libra_types::proptest_types

Structs

AccountInfoUniverse

ContractEventGen

LedgerInfoWithSignaturesGen

RawTransactionGen

SignatureCheckedTransactionGen

TransactionToCommitGen
Represents information already determined for generating a TransactionToCommit, along with to be determined information that needs to settle upon materialize(), for example a to be determined account can be represented by an Index which will be materialized to an entry in the AccountInfoUniverse.