Module libra_types::proto::types

Modules

request_item

response_item

transaction_argument

Structs

AccessPath

AccountState
Account state as a whole. After execution, updates to accounts are passed in this form to storage for persistence.

AccountStateBlob

AccountStateProof
The complete proof used to authenticate an account state.

AccountStateWithProof

AccumulatorConsistencyProof

AccumulatorProof

AccumulatorRangeProof

EpochChangeProof
This is used to prove epoch changes.

Event
An event emitted from a smart contract

EventProof
The complete proof used to authenticate an event.

EventWithProof
An event along with the proof for the event

EventsForVersions
A list of EventList's, each representing all events for a transaction.

EventsList
A list of events.

GetAccountStateRequest
Gets latest state for an account.

GetAccountStateResponse
State information returned by a get account state query.

GetAccountTransactionBySequenceNumberRequest

GetAccountTransactionBySequenceNumberResponse
Transaction information for transactions requested by GetAccountTransactionsRequest

GetEventsByEventAccessPathRequest
Get events that exist on an event access path. In the current world, a user may specify events that were received, events that were sent, or any event that modifies their account

GetEventsByEventAccessPathResponse

GetTransactionsRequest
Get up to limit transactions starting from start_version.

GetTransactionsResponse

LedgerInfo

LedgerInfoWithSignatures

MempoolStatus
The statuses and errors produced by the Mempool during transaction insertion

RequestItem

ResponseItem
Individual response items to the queries posed by the requests

SignedTransaction
A generic structure that represents signed RawTransaction

SignedTransactionsBlock
A generic structure that represents a block of transactions originated from a particular validator instance.

SparseMerkleProof

SparseMerkleRangeProof

Transaction
A generic structure that represents a transaction, covering all possible variants.

TransactionArgument
An argument to the transaction if the transaction takes arguments

TransactionInfo
TransactionInfo is the object we store in the transaction accumulator. It consists of the transaction as well as the execution result of this transaction. This are later returned to the client so that a client can validate the tree

TransactionListProof
The complete proof used to authenticate a list of transactions.

TransactionListWithProof
A list of consecutive transactions with proof. This is mainly used for state synchronization when a validator would request a list of transactions from a peer, verify the proof, execute the transactions and persist them. Note that the transactions are supposed to belong to the same epoch E, otherwise verification will fail.

TransactionProof
The complete proof used to authenticate a transaction.

TransactionToCommit
Transaction struct to commit to storage

TransactionWithProof

UpdateToLatestLedgerRequest
This API is used to update the client to the latest ledger version and optionally also request 1..n other pieces of data. This allows for batch queries. All queries return proofs that a client should check to validate the data.

UpdateToLatestLedgerResponse
Response from getting latest ledger

ValidatorInfo
Protobuf definition for the Rust struct ValidatorInfo

ValidatorSet
Protobuf definition for the Rust struct ValidatorSet. Serialized by LCS

ValidatorSignature

VmStatus
The statuses and errors produced by the VM can be categorized into a couple different types: