Enum libra_types::transaction::Transaction

pub enum Transaction {
    UserTransaction(SignedTransaction),
    WaypointWriteSet(ChangeSet),
    BlockMetadata(BlockMetadata),
}

Transaction will be the transaction type used internally in the libra node to represent the transaction to be processed and persisted.

We suppress the clippy warning here as we would expect most of the transaction to be user transaction.

Variants

UserTransaction(SignedTransaction)

Transaction submitted by the user. e.g: P2P payment transaction, publishing module transaction, etc. TODO: We need to rename SignedTransaction to SignedUserTransaction, as well as all the other transaction types we had in our codebase.

WaypointWriteSet(ChangeSet)

Transaction that applies a WriteSet to the current storage. This should be used for ONLY for genesis right now.

BlockMetadata(BlockMetadata)

Transaction to update the block metadata resource at the beginning of a block.

Methods

impl Transaction

pub fn as_signed_user_txn(&self) -> Result<&SignedTransaction>

pub fn format_for_client(
&self,
get_transaction_name: impl Fn(&[u8]) -> String
) -> String