Module libra_types::vm_error

Modules

sub_status

Structs

VMStatus
A VMStatus is represented as a required major status that is semantic coupled with with an optional sub status and message.

Enums

StatusCode
We don't derive Arbitrary on this enum because it is too large and breaks proptest. It is written for a subset of these in proptest_types. We test conversion between this and protobuf with a hand-written test.

StatusType
A status type is one of 5 different variants, along with a fallback variant in the case that we don't recognize the status code.

Statics

DESERIALIZATION_STATUS_MAX_CODE
The maximum status code for deserialization statuses

DESERIALIZATION_STATUS_MIN_CODE
The minimum status code for deserialization statuses

EXECUTION_STATUS_MAX_CODE
The maximum status code for runtim statuses

EXECUTION_STATUS_MIN_CODE
The minimum status code for runtime statuses

INVARIANT_VIOLATION_STATUS_MAX_CODE
The maximum status code for invariant violation statuses

INVARIANT_VIOLATION_STATUS_MIN_CODE
The minimum status code for invariant violation statuses

VALIDATION_STATUS_MAX_CODE
The maximum status code for validation statuses

VALIDATION_STATUS_MIN_CODE
The minimum status code for validation statuses

VERIFICATION_STATUS_MAX_CODE
The maximum status code for verification statuses

VERIFICATION_STATUS_MIN_CODE
The minimum status code for verification statuses