Enum libra_types::vm_error::StatusCode

#[repr(u64)]
pub enum StatusCode {
    UNKNOWN_VALIDATION_STATUS,
    INVALID_SIGNATURE,
    INVALID_AUTH_KEY,
    SEQUENCE_NUMBER_TOO_OLD,
    SEQUENCE_NUMBER_TOO_NEW,
    INSUFFICIENT_BALANCE_FOR_TRANSACTION_FEE,
    TRANSACTION_EXPIRED,
    SENDING_ACCOUNT_DOES_NOT_EXIST,
    REJECTED_WRITE_SET,
    INVALID_WRITE_SET,
    EXCEEDED_MAX_TRANSACTION_SIZE,
    UNKNOWN_SCRIPT,
    UNKNOWN_MODULE,
    MAX_GAS_UNITS_EXCEEDS_MAX_GAS_UNITS_BOUND,
    MAX_GAS_UNITS_BELOW_MIN_TRANSACTION_GAS_UNITS,
    GAS_UNIT_PRICE_BELOW_MIN_BOUND,
    GAS_UNIT_PRICE_ABOVE_MAX_BOUND,
    INVALID_GAS_SPECIFIER,
    SENDING_ACCOUNT_FROZEN,
    UNABLE_TO_DESERIALIZE_ACCOUNT,
    CURRENCY_INFO_DOES_NOT_EXIST,
    UNKNOWN_VERIFICATION_ERROR,
    INDEX_OUT_OF_BOUNDS,
    RANGE_OUT_OF_BOUNDS,
    INVALID_SIGNATURE_TOKEN,
    INVALID_FIELD_DEF,
    RECURSIVE_STRUCT_DEFINITION,
    INVALID_RESOURCE_FIELD,
    INVALID_FALL_THROUGH,
    JOIN_FAILURE,
    NEGATIVE_STACK_SIZE_WITHIN_BLOCK,
    UNBALANCED_STACK,
    INVALID_MAIN_FUNCTION_SIGNATURE,
    DUPLICATE_ELEMENT,
    INVALID_MODULE_HANDLE,
    UNIMPLEMENTED_HANDLE,
    INCONSISTENT_FIELDS,
    UNUSED_FIELD,
    LOOKUP_FAILED,
    VISIBILITY_MISMATCH,
    TYPE_RESOLUTION_FAILURE,
    TYPE_MISMATCH,
    MISSING_DEPENDENCY,
    POP_REFERENCE_ERROR,
    POP_RESOURCE_ERROR,
    RELEASEREF_TYPE_MISMATCH_ERROR,
    BR_TYPE_MISMATCH_ERROR,
    ABORT_TYPE_MISMATCH_ERROR,
    STLOC_TYPE_MISMATCH_ERROR,
    STLOC_UNSAFE_TO_DESTROY_ERROR,
    UNSAFE_RET_LOCAL_OR_RESOURCE_STILL_BORROWED,
    RET_TYPE_MISMATCH_ERROR,
    RET_BORROWED_MUTABLE_REFERENCE_ERROR,
    FREEZEREF_TYPE_MISMATCH_ERROR,
    FREEZEREF_EXISTS_MUTABLE_BORROW_ERROR,
    BORROWFIELD_TYPE_MISMATCH_ERROR,
    BORROWFIELD_BAD_FIELD_ERROR,
    BORROWFIELD_EXISTS_MUTABLE_BORROW_ERROR,
    COPYLOC_UNAVAILABLE_ERROR,
    COPYLOC_RESOURCE_ERROR,
    COPYLOC_EXISTS_BORROW_ERROR,
    MOVELOC_UNAVAILABLE_ERROR,
    MOVELOC_EXISTS_BORROW_ERROR,
    BORROWLOC_REFERENCE_ERROR,
    BORROWLOC_UNAVAILABLE_ERROR,
    BORROWLOC_EXISTS_BORROW_ERROR,
    CALL_TYPE_MISMATCH_ERROR,
    CALL_BORROWED_MUTABLE_REFERENCE_ERROR,
    PACK_TYPE_MISMATCH_ERROR,
    UNPACK_TYPE_MISMATCH_ERROR,
    READREF_TYPE_MISMATCH_ERROR,
    READREF_RESOURCE_ERROR,
    READREF_EXISTS_MUTABLE_BORROW_ERROR,
    WRITEREF_TYPE_MISMATCH_ERROR,
    WRITEREF_RESOURCE_ERROR,
    WRITEREF_EXISTS_BORROW_ERROR,
    WRITEREF_NO_MUTABLE_REFERENCE_ERROR,
    INTEGER_OP_TYPE_MISMATCH_ERROR,
    BOOLEAN_OP_TYPE_MISMATCH_ERROR,
    EQUALITY_OP_TYPE_MISMATCH_ERROR,
    EXISTS_RESOURCE_TYPE_MISMATCH_ERROR,
    BORROWGLOBAL_TYPE_MISMATCH_ERROR,
    BORROWGLOBAL_NO_RESOURCE_ERROR,
    MOVEFROM_TYPE_MISMATCH_ERROR,
    MOVEFROM_NO_RESOURCE_ERROR,
    MOVETOSENDER_TYPE_MISMATCH_ERROR,
    MOVETOSENDER_NO_RESOURCE_ERROR,
    CREATEACCOUNT_TYPE_MISMATCH_ERROR,
    MODULE_ADDRESS_DOES_NOT_MATCH_SENDER,
    NO_MODULE_HANDLES,
    POSITIVE_STACK_SIZE_AT_BLOCK_END,
    MISSING_ACQUIRES_RESOURCE_ANNOTATION_ERROR,
    EXTRANEOUS_ACQUIRES_RESOURCE_ANNOTATION_ERROR,
    DUPLICATE_ACQUIRES_RESOURCE_ANNOTATION_ERROR,
    INVALID_ACQUIRES_RESOURCE_ANNOTATION_ERROR,
    GLOBAL_REFERENCE_ERROR,
    CONTRAINT_KIND_MISMATCH,
    NUMBER_OF_TYPE_ARGUMENTS_MISMATCH,
    LOOP_IN_INSTANTIATION_GRAPH,
    UNUSED_LOCALS_SIGNATURE,
    UNUSED_TYPE_SIGNATURE,
    ZERO_SIZED_STRUCT,
    LINKER_ERROR,
    INVALID_CONSTANT_TYPE,
    MALFORMED_CONSTANT_DATA,
    EMPTY_CODE_UNIT,
    INVALID_LOOP_SPLIT,
    INVALID_LOOP_BREAK,
    INVALID_LOOP_CONTINUE,
    UNSAFE_RET_UNUSED_RESOURCES,
    TOO_MANY_LOCALS,
    MOVETO_TYPE_MISMATCH_ERROR,
    MOVETO_NO_RESOURCE_ERROR,
    UNKNOWN_INVARIANT_VIOLATION_ERROR,
    OUT_OF_BOUNDS_INDEX,
    OUT_OF_BOUNDS_RANGE,
    EMPTY_VALUE_STACK,
    EMPTY_CALL_STACK,
    PC_OVERFLOW,
    VERIFICATION_ERROR,
    LOCAL_REFERENCE_ERROR,
    STORAGE_ERROR,
    INTERNAL_TYPE_ERROR,
    EVENT_KEY_MISMATCH,
    UNREACHABLE,
    VM_STARTUP_FAILURE,
    NATIVE_FUNCTION_INTERNAL_INCONSISTENCY,
    INVALID_CODE_CACHE,
    UNKNOWN_BINARY_ERROR,
    MALFORMED,
    BAD_MAGIC,
    UNKNOWN_VERSION,
    UNKNOWN_TABLE_TYPE,
    UNKNOWN_SIGNATURE_TYPE,
    UNKNOWN_SERIALIZED_TYPE,
    UNKNOWN_OPCODE,
    BAD_HEADER_TABLE,
    UNEXPECTED_SIGNATURE_TYPE,
    DUPLICATE_TABLE,
    VERIFIER_INVARIANT_VIOLATION,
    UNKNOWN_NOMINAL_RESOURCE,
    UNKNOWN_KIND,
    UNKNOWN_NATIVE_STRUCT_FLAG,
    BAD_ULEB_U16,
    BAD_ULEB_U32,
    BAD_U16,
    BAD_U32,
    BAD_U64,
    BAD_U128,
    UNKNOWN_RUNTIME_STATUS,
    EXECUTED,
    OUT_OF_GAS,
    RESOURCE_DOES_NOT_EXIST,
    RESOURCE_ALREADY_EXISTS,
    EVICTED_ACCOUNT_ACCESS,
    ACCOUNT_ADDRESS_ALREADY_EXISTS,
    TYPE_ERROR,
    MISSING_DATA,
    DATA_FORMAT_ERROR,
    INVALID_DATA,
    REMOTE_DATA_ERROR,
    CANNOT_WRITE_EXISTING_RESOURCE,
    VALUE_SERIALIZATION_ERROR,
    VALUE_DESERIALIZATION_ERROR,
    DUPLICATE_MODULE_NAME,
    ABORTED,
    ARITHMETIC_ERROR,
    DYNAMIC_REFERENCE_ERROR,
    CODE_DESERIALIZATION_ERROR,
    EXECUTION_STACK_OVERFLOW,
    CALL_STACK_OVERFLOW,
    NATIVE_FUNCTION_ERROR,
    GAS_SCHEDULE_ERROR,
    UNKNOWN_STATUS,
}

We don't derive Arbitrary on this enum because it is too large and breaks proptest. It is written for a subset of these in proptest_types. We test conversion between this and protobuf with a hand-written test.

Variants

UNKNOWN_VALIDATION_STATUS

INVALID_SIGNATURE

INVALID_AUTH_KEY

SEQUENCE_NUMBER_TOO_OLD

SEQUENCE_NUMBER_TOO_NEW

INSUFFICIENT_BALANCE_FOR_TRANSACTION_FEE

TRANSACTION_EXPIRED

SENDING_ACCOUNT_DOES_NOT_EXIST

REJECTED_WRITE_SET

INVALID_WRITE_SET

EXCEEDED_MAX_TRANSACTION_SIZE

UNKNOWN_SCRIPT

UNKNOWN_MODULE

MAX_GAS_UNITS_EXCEEDS_MAX_GAS_UNITS_BOUND

MAX_GAS_UNITS_BELOW_MIN_TRANSACTION_GAS_UNITS

GAS_UNIT_PRICE_BELOW_MIN_BOUND

GAS_UNIT_PRICE_ABOVE_MAX_BOUND

INVALID_GAS_SPECIFIER

SENDING_ACCOUNT_FROZEN

UNABLE_TO_DESERIALIZE_ACCOUNT

CURRENCY_INFO_DOES_NOT_EXIST

UNKNOWN_VERIFICATION_ERROR

INDEX_OUT_OF_BOUNDS

RANGE_OUT_OF_BOUNDS

INVALID_SIGNATURE_TOKEN

INVALID_FIELD_DEF

RECURSIVE_STRUCT_DEFINITION

INVALID_RESOURCE_FIELD

INVALID_FALL_THROUGH

JOIN_FAILURE

NEGATIVE_STACK_SIZE_WITHIN_BLOCK

UNBALANCED_STACK

INVALID_MAIN_FUNCTION_SIGNATURE

DUPLICATE_ELEMENT

INVALID_MODULE_HANDLE

UNIMPLEMENTED_HANDLE

INCONSISTENT_FIELDS

UNUSED_FIELD

LOOKUP_FAILED

VISIBILITY_MISMATCH

TYPE_RESOLUTION_FAILURE

TYPE_MISMATCH

MISSING_DEPENDENCY

POP_REFERENCE_ERROR

POP_RESOURCE_ERROR

RELEASEREF_TYPE_MISMATCH_ERROR

BR_TYPE_MISMATCH_ERROR

ABORT_TYPE_MISMATCH_ERROR

STLOC_TYPE_MISMATCH_ERROR

STLOC_UNSAFE_TO_DESTROY_ERROR

UNSAFE_RET_LOCAL_OR_RESOURCE_STILL_BORROWED

RET_TYPE_MISMATCH_ERROR

RET_BORROWED_MUTABLE_REFERENCE_ERROR

FREEZEREF_TYPE_MISMATCH_ERROR

FREEZEREF_EXISTS_MUTABLE_BORROW_ERROR

BORROWFIELD_TYPE_MISMATCH_ERROR

BORROWFIELD_BAD_FIELD_ERROR

BORROWFIELD_EXISTS_MUTABLE_BORROW_ERROR

COPYLOC_UNAVAILABLE_ERROR

COPYLOC_RESOURCE_ERROR

COPYLOC_EXISTS_BORROW_ERROR

MOVELOC_UNAVAILABLE_ERROR

MOVELOC_EXISTS_BORROW_ERROR

BORROWLOC_REFERENCE_ERROR

BORROWLOC_UNAVAILABLE_ERROR

BORROWLOC_EXISTS_BORROW_ERROR

CALL_TYPE_MISMATCH_ERROR

CALL_BORROWED_MUTABLE_REFERENCE_ERROR

PACK_TYPE_MISMATCH_ERROR

UNPACK_TYPE_MISMATCH_ERROR

READREF_TYPE_MISMATCH_ERROR

READREF_RESOURCE_ERROR

READREF_EXISTS_MUTABLE_BORROW_ERROR

WRITEREF_TYPE_MISMATCH_ERROR

WRITEREF_RESOURCE_ERROR

WRITEREF_EXISTS_BORROW_ERROR

WRITEREF_NO_MUTABLE_REFERENCE_ERROR

INTEGER_OP_TYPE_MISMATCH_ERROR

BOOLEAN_OP_TYPE_MISMATCH_ERROR

EQUALITY_OP_TYPE_MISMATCH_ERROR

EXISTS_RESOURCE_TYPE_MISMATCH_ERROR

BORROWGLOBAL_TYPE_MISMATCH_ERROR

BORROWGLOBAL_NO_RESOURCE_ERROR

MOVEFROM_TYPE_MISMATCH_ERROR

MOVEFROM_NO_RESOURCE_ERROR

MOVETOSENDER_TYPE_MISMATCH_ERROR

MOVETOSENDER_NO_RESOURCE_ERROR

CREATEACCOUNT_TYPE_MISMATCH_ERROR

MODULE_ADDRESS_DOES_NOT_MATCH_SENDER

NO_MODULE_HANDLES

POSITIVE_STACK_SIZE_AT_BLOCK_END

MISSING_ACQUIRES_RESOURCE_ANNOTATION_ERROR

EXTRANEOUS_ACQUIRES_RESOURCE_ANNOTATION_ERROR

DUPLICATE_ACQUIRES_RESOURCE_ANNOTATION_ERROR

INVALID_ACQUIRES_RESOURCE_ANNOTATION_ERROR

GLOBAL_REFERENCE_ERROR

CONTRAINT_KIND_MISMATCH

NUMBER_OF_TYPE_ARGUMENTS_MISMATCH

LOOP_IN_INSTANTIATION_GRAPH

UNUSED_LOCALS_SIGNATURE

UNUSED_TYPE_SIGNATURE

ZERO_SIZED_STRUCT

Reported when a struct has zero fields

LINKER_ERROR

INVALID_CONSTANT_TYPE

MALFORMED_CONSTANT_DATA

EMPTY_CODE_UNIT

INVALID_LOOP_SPLIT

INVALID_LOOP_BREAK

INVALID_LOOP_CONTINUE

UNSAFE_RET_UNUSED_RESOURCES

TOO_MANY_LOCALS

MOVETO_TYPE_MISMATCH_ERROR

MOVETO_NO_RESOURCE_ERROR

UNKNOWN_INVARIANT_VIOLATION_ERROR

OUT_OF_BOUNDS_INDEX

OUT_OF_BOUNDS_RANGE

EMPTY_VALUE_STACK

EMPTY_CALL_STACK

PC_OVERFLOW

VERIFICATION_ERROR

LOCAL_REFERENCE_ERROR

STORAGE_ERROR

INTERNAL_TYPE_ERROR

EVENT_KEY_MISMATCH

UNREACHABLE

VM_STARTUP_FAILURE

NATIVE_FUNCTION_INTERNAL_INCONSISTENCY

INVALID_CODE_CACHE

UNKNOWN_BINARY_ERROR

MALFORMED

BAD_MAGIC

UNKNOWN_VERSION

UNKNOWN_TABLE_TYPE

UNKNOWN_SIGNATURE_TYPE

UNKNOWN_SERIALIZED_TYPE

UNKNOWN_OPCODE

BAD_HEADER_TABLE

UNEXPECTED_SIGNATURE_TYPE

DUPLICATE_TABLE

VERIFIER_INVARIANT_VIOLATION

UNKNOWN_NOMINAL_RESOURCE

UNKNOWN_KIND

UNKNOWN_NATIVE_STRUCT_FLAG

BAD_ULEB_U16

BAD_ULEB_U32

BAD_U16

BAD_U32

BAD_U64

BAD_U128

UNKNOWN_RUNTIME_STATUS

EXECUTED

OUT_OF_GAS

RESOURCE_DOES_NOT_EXIST

RESOURCE_ALREADY_EXISTS

EVICTED_ACCOUNT_ACCESS

ACCOUNT_ADDRESS_ALREADY_EXISTS

TYPE_ERROR

MISSING_DATA

DATA_FORMAT_ERROR

INVALID_DATA

REMOTE_DATA_ERROR

CANNOT_WRITE_EXISTING_RESOURCE

VALUE_SERIALIZATION_ERROR

VALUE_DESERIALIZATION_ERROR

DUPLICATE_MODULE_NAME

ABORTED

ARITHMETIC_ERROR

DYNAMIC_REFERENCE_ERROR

CODE_DESERIALIZATION_ERROR

EXECUTION_STACK_OVERFLOW

CALL_STACK_OVERFLOW

NATIVE_FUNCTION_ERROR

GAS_SCHEDULE_ERROR

UNKNOWN_STATUS