Struct libra_types::vm_error::VMStatus

pub struct VMStatus {
    pub major_status: StatusCode,
    pub sub_status: Option<u64>,
    pub message: Option<String>,
}

A VMStatus is represented as a required major status that is semantic coupled with with an optional sub status and message.

Fields
major_status: StatusCode

The major status, e.g. ABORTED, OUT_OF_GAS, etc.

sub_status: Option<u64>

The optional sub status. Used e.g. for things such as the abort code, or the arithmetic error type for an ARITHMETIC_ERROR major status.

message: Option<String>

The optional message. Useful for verification errors, and for returning information in validation.

Methods

impl VMStatus

pub fn new(major_status: StatusCode) -> Self

Create a new VM status with major status major_status.

pub fn with_sub_status(self, sub_status: u64) -> Self

Adds a sub status to the VM status.

pub fn with_message(self, message: String) -> Self

Adds a message to the VM status.

pub fn set_sub_status(&mut self, sub_status: u64)

Mutates the VMStatus sub status field to be the new sub_status passed in.

pub fn set_message(&mut self, message: String)

Mutates the VMStatus message field to be the new message passed in.

pub fn append_message_with_separator(
self,
separator: char,
message: String
) -> Self

Append the message message to the message field of the VM status, and insert a seperator if the original message is non-empty.

pub fn status_type(&self) -> StatusType

Return the status type for this VMStatus. This is solely determined by the major_status field.

pub fn is(&self, status_type: StatusType) -> bool

Determine if the VMStatus has status type status_type.

pub fn append(self, other: Self) -> Self

Append two VMStatuses together. The major status is kept from the caller.