Module libra_types::write_set

For each transaction the VM executes, the VM will output a WriteSet that contains each access path it updates. For each access path, the VM can either give its new value or delete it.

Structs

WriteSet
WriteSet contains all access paths that one transaction modifies. Each of them is a WriteOp where Value(val) means that serialized representation should be updated to val, and Deletion means that we are going to delete this access path.

WriteSetMut
A mutable version of WriteSet.

Enums

WriteOp