Crate libra_types

Modules

access_path
Suppose we have the following data structure in a smart contract:

account_address

account_config

account_state

account_state_blob

block_info

block_metadata

contract_event

discovery_info

discovery_set

epoch_change

epoch_info

event

get_with_proof

language_storage

ledger_info

libra_timestamp

mempool_status

move_resource

on_chain_config

proof

proptest_types

proto

test_helpers

transaction

trusted_state

validator_config

validator_info

validator_signer

validator_verifier

vm_error

waypoint

write_set
For each transaction the VM executes, the VM will output a WriteSet that contains each access path it updates. For each access path, the VM can either give its new value or delete it.

Structs

PeerId
A struct that represents an account address.